jason/20251120_timescale_correct_vs_incorrect/requirements.txt
boto3==1.41.1
botocore==1.41.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
colorcet==3.1.0
colorlog==6.10.1
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
dask==2025.11.0
-e git+ssh://git@github.com/kbstafford/decisivetimes.git@acf524a4a1b83eb9edbe212f5a3f712e7ead6ee6#egg=decisive_times_utils&subdirectory=decisive-times-utils
figrid==0.1.7
flake8==7.3.0
fonttools==4.60.1
fsspec==2025.10.0
globus-sdk==4.1.0
graphviz==0.21
ibl-neuropixel==1.9.1
ibl-style==0.1.0
iblatlas==0.10.0
ibllib==3.4.1
iblqt==0.8.2
iblutil==1.20.0
idna==3.11
imagecodecs==2025.11.11
ImageIO==2.37.2
importlib_metadata==8.7.0
iniconfig==2.3.0
jmespath==1.0.1
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
llvmlite==0.45.1
locket==1.0.0
matplotlib==3.10.7
mccabe==0.7.0
mtscomp==1.0.2
networkx==3.5
npTDMS==1.10.0
numba==0.62.1
numpy==2.2.0
ONE-api==3.4.1
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.3
partd==1.4.2
patsy==1.0.2
phylib==2.6.3
pillow==12.0.0
pluggy==1.6.0
Psychofit==1.0.0.post0
pyarrow==22.0.0
pycodestyle==2.14.0
pycparser==2.23
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pynrrd==1.1.3
pyparsing==3.2.5
PyQt5==5.15.11
PyQt5-Qt5==5.15.18
PyQt5_sip==12.17.1
pyqtgraph==0.14.0
pytest==9.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
QtPy==2.4.3
requests==2.32.5
responses==0.25.8
ruff==0.14.5
s3transfer==0.15.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
six==1.17.0
slidingRP==1.1.1
sparse==0.17.0
statsmodels==0.14.5
threadpoolctl==3.6.0
tifffile==2025.10.16
toolz==1.1.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
zipp==3.23.0
